+++ packages/core/src/GlobalStyles/GlobalStyles.tsx
@@ -90,7 +90,6 @@ export const GlobalStyles = ({
           border: 0;
           box-sizing: border-box;
           margin: 0;
-          min-height: 100%;
           padding: 0;
         }