+++ packages/components/icons/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.1.0](https://github.com/contentful/forma-36/compare/@contentful/f36-icons@4.0.4...@contentful/f36-icons@4.1.0) (2022-02-09)
+
+
+### Features
+
+* add new icon - arrow forward ([#1867](https://github.com/contentful/forma-36/issues/1867)) ([cc36893](https://github.com/contentful/forma-36/commit/cc368932b8271f7fefcd194cf35d88f7fae90383))
+
+
+
+
+
+## [4.0.4](https://github.com/contentful/forma-36/compare/@contentful/f36-icons@4.0.3...@contentful/f36-icons@4.0.4) (2022-02-07)
+
+
+### Bug Fixes
+
+* add preview off icon ([#1866](https://github.com/contentful/forma-36/issues/1866)) ([d88a518](https://github.com/contentful/forma-36/commit/d88a518a44e5e1a10fe3a695e4e0824e472632dc))
+
+
+
+
+
+## [4.0.3](https://github.com/contentful/forma-36/compare/@contentful/f36-icons@4.0.2...@contentful/f36-icons@4.0.3) (2022-01-31)
+
+**Note:** Version bump only for package @contentful/f36-icons
+
+
+
+
+
+## [4.0.2](https://github.com/contentful/forma-36/compare/@contentful/f36-icons@4.0.1...@contentful/f36-icons@4.0.2) (2022-01-31)
+
+**Note:** Version bump only for package @contentful/f36-icons
+
+
+
+
+
 ## [4.0.1](https://github.com/contentful/forma-36/compare/@contentful/f36-icons@4.0.0...@contentful/f36-icons@4.0.1) (2022-01-11)
 
 **Note:** Version bump only for package @contentful/f36-icons