+++ packages/components/modal/CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.1](https://github.com/contentful/forma-36/compare/@contentful/f36-modal@4.2.0...@contentful/f36-modal@4.2.1) (2022-02-09)
+
+**Note:** Version bump only for package @contentful/f36-modal
+
+
+
+
+
+# [4.2.0](https://github.com/contentful/forma-36/compare/@contentful/f36-modal@4.1.8...@contentful/f36-modal@4.2.0) (2022-02-07)
+
+
+### Features
+
+* implement short description in page content header ([#1822](https://github.com/contentful/forma-36/issues/1822)) ([c812815](https://github.com/contentful/forma-36/commit/c812815cf6681a021fd5313fc905227def842ee8))
+
+
+
+
+
+## [4.1.8](https://github.com/contentful/forma-36/compare/@contentful/f36-modal@4.1.7...@contentful/f36-modal@4.1.8) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-modal
+
+
+
+
+
+## [4.1.7](https://github.com/contentful/forma-36/compare/@contentful/f36-modal@4.1.6...@contentful/f36-modal@4.1.7) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-modal
+
+
+
+
+
+## [4.1.6](https://github.com/contentful/forma-36/compare/@contentful/f36-modal@4.1.5...@contentful/f36-modal@4.1.6) (2022-02-02)
+
+**Note:** Version bump only for package @contentful/f36-modal
+
+
+
+
+
+## [4.1.5](https://github.com/contentful/forma-36/compare/@contentful/f36-modal@4.1.4...@contentful/f36-modal@4.1.5) (2022-01-31)
+
+**Note:** Version bump only for package @contentful/f36-modal
+
+
+
+
+
+## [4.1.4](https://github.com/contentful/forma-36/compare/@contentful/f36-modal@4.1.3...@contentful/f36-modal@4.1.4) (2022-01-31)
+
+**Note:** Version bump only for package @contentful/f36-modal
+
+
+
+
+
+## [4.1.3](https://github.com/contentful/forma-36/compare/@contentful/f36-modal@4.1.2...@contentful/f36-modal@4.1.3) (2022-01-27)
+
+
+### Bug Fixes
+
+* fix Modal defaults, adjust docs ([#1827](https://github.com/contentful/forma-36/issues/1827)) ([7a262f0](https://github.com/contentful/forma-36/commit/7a262f0719492fba16e0557e7fc3316eeacd2a21))
+
+
+
+
+
 ## [4.1.2](https://github.com/contentful/forma-36/compare/@contentful/f36-modal@4.1.1...@contentful/f36-modal@4.1.2) (2022-01-26)
 
 **Note:** Version bump only for package @contentful/f36-modal