+++ packages/components/notification/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.11](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.10...@contentful/f36-notification@4.0.11) (2022-02-09)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
+## [4.0.10](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.9...@contentful/f36-notification@4.0.10) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
+## [4.0.9](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.8...@contentful/f36-notification@4.0.9) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
+## [4.0.8](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.7...@contentful/f36-notification@4.0.8) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
+## [4.0.7](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.6...@contentful/f36-notification@4.0.7) (2022-02-02)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
+## [4.0.6](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.5...@contentful/f36-notification@4.0.6) (2022-01-31)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
+## [4.0.5](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.4...@contentful/f36-notification@4.0.5) (2022-01-31)
+
+**Note:** Version bump only for package @contentful/f36-notification
+
+
+
+
+
 ## [4.0.4](https://github.com/contentful/forma-36/compare/@contentful/f36-notification@4.0.3...@contentful/f36-notification@4.0.4) (2022-01-26)
 
 **Note:** Version bump only for package @contentful/f36-notification