+++ packages/core/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.1](https://github.com/contentful/forma-36/compare/@contentful/f36-core@4.1.0...@contentful/f36-core@4.1.1) (2022-01-31)
+
+
+### Bug Fixes
+
+* remove min-height from GlobalStyles ([#1843](https://github.com/contentful/forma-36/issues/1843)) ([dfff19d](https://github.com/contentful/forma-36/commit/dfff19dc1b502bb1c325dd6555c1fe89a41ea748))
+
+
+
+
+
+# [4.1.0](https://github.com/contentful/forma-36/compare/@contentful/f36-core@4.0.1...@contentful/f36-core@4.1.0) (2022-01-31)
+
+
+### Bug Fixes
+
+* **core:** disable ExpandProps while we investigate a bug caused by it ([#1836](https://github.com/contentful/forma-36/issues/1836)) ([35e30ce](https://github.com/contentful/forma-36/commit/35e30ceafd8c813a8775d339d6a6631c2d0b6011))
+
+
+### Features
+
+* implement props list redesign [BAU-535]  ([#1756](https://github.com/contentful/forma-36/issues/1756)) ([21c57e7](https://github.com/contentful/forma-36/commit/21c57e72008b75990d03af4e7500edc1c7f3d26d))
+
+
+
+
+
 ## [4.0.1](https://github.com/contentful/forma-36/compare/@contentful/f36-core@4.0.0...@contentful/f36-core@4.0.1) (2022-01-11)
 
 **Note:** Version bump only for package @contentful/f36-core