+++ packages/forma-36-react-components/CHANGELOG.md
@@ -3,6 +3,142 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.29](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.28...@contentful/f36-components@4.0.29) (2022-02-15)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.28](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.27...@contentful/f36-components@4.0.28) (2022-02-15)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.27](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.26...@contentful/f36-components@4.0.27) (2022-02-14)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.26](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.25...@contentful/f36-components@4.0.26) (2022-02-09)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.25](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.24...@contentful/f36-components@4.0.25) (2022-02-09)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.24](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.23...@contentful/f36-components@4.0.24) (2022-02-09)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.23](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.22...@contentful/f36-components@4.0.23) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.22](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.21...@contentful/f36-components@4.0.22) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.21](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.20...@contentful/f36-components@4.0.21) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.20](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.19...@contentful/f36-components@4.0.20) (2022-02-04)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.19](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.18...@contentful/f36-components@4.0.19) (2022-02-02)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.18](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.17...@contentful/f36-components@4.0.18) (2022-02-02)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.17](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.16...@contentful/f36-components@4.0.17) (2022-01-31)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.16](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.15...@contentful/f36-components@4.0.16) (2022-01-31)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.15](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.14...@contentful/f36-components@4.0.15) (2022-01-28)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.14](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.13...@contentful/f36-components@4.0.14) (2022-01-28)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.13](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.12...@contentful/f36-components@4.0.13) (2022-01-27)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
 ## [4.0.12](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.11...@contentful/f36-components@4.0.12) (2022-01-26)
 
 **Note:** Version bump only for package @contentful/f36-components